# docs/swagger.yaml
definitions:
...
  models.Product:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_urls:
        items:
          type: string
        type: array
      ingredients:
        type: string
      name:
        type: string
      nutritional_info:
        type: string
      updated_at:
        type: string
      weights:
        items:
          $ref: '#/definitions/models.ProductWeight'
        type: array
    type: object
  models.ProductWeight:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      price:
        type: number
      productID:
        type: integer
      stockAvailability:
        type: integer
      weight:
        type: integer
    type: object
...
  services.AddProductRequest:
    properties:
      category:
        type: string
      description:
        type: string
      image_urls:
        items:
          type: string
        type: array
      ingredients:
        type: string
      name:
        type: string
      nutritional_info:
        type: string
    type: object
  services.AddProductResponse:
    properties:
      message:
        type: string
      product_id:
        type: integer
    type: object
...
  services.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Failed to retrieve products
      summary: List all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/services.AddProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product added successfully
          schema:
            $ref: '#/definitions/services.AddProductResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/services.ErrorResponse'
        "500":
          description: Failed to add product
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Add a new product to the inventory
      tags:
      - Products
  /products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Product deleted successfully
        "500":
          description: Failed to delete product
      summary: Delete a product from the inventory
      tags:
      - Products
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/models.Product'
        "500":
          description: Failed to retrieve product details
      summary: Get product details by ID
      tags:
      - Products
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product name
        in: formData
        name: name
        required: true
        type: string
      - description: Product description
        in: formData
        name: description
        required: true
        type: string
      - description: Product category
        in: formData
        name: category
        required: true
        type: string
      - description: Product ingredients
        in: formData
        name: ingredients
        required: true
        type: string
      - description: Product nutritional information
        in: formData
        name: nutritional_info
        required: true
        type: string
      - description: Product image URLs (comma-separated)
        in: formData
        name: image_urls
        required: true
        type: string
      responses:
        "200":
          description: Product updated successfully
        "400":
          description: Invalid form data
        "500":
          description: Failed to update product
      summary: Update an existing product in the inventory
      tags:
      - Products
...
swagger: "2.0"